+++ test/test_parse.c
@@ -6759,6 +6759,37 @@ START_TEST(test_issue942)
 }
 END_TEST
 
+START_TEST(test_lcs2016_i03)
+{
+   set_standard(STD_19);
+
+   input_from_file(TESTDIR "/parse/lcs2016_i03.vhd");
+
+   const error_t expect[] = {
+      { 27, "a signature is only allowed in a generic formal designator" },
+      { 33, "no visible subprogram or enumeration literal F1 matches "
+        "signature [return BOOLEAN]" },
+      { 39, "no visible subprogram or enumeration literal G matches "
+        "signature [return INTEGER]" },
+      { -1, NULL }
+   };
+   expect_errors(expect);
+
+   tree_t e = parse();
+   fail_if(e == NULL);
+   fail_unless(tree_kind(e) == T_ENTITY);
+   lib_put(lib_work(), e);
+
+   tree_t a = parse();
+   fail_if(a == NULL);
+   fail_unless(tree_kind(a) == T_ARCH);
+
+   fail_unless(parse() == NULL);
+
+   check_expected_errors();
+}
+END_TEST
+
 Suite *get_parse_tests(void)
 {
    Suite *s = suite_create("parse");
@@ -6916,6 +6947,7 @@ Suite *get_parse_tests(void)
    tcase_add_test(tc_core, test_issue905);
    tcase_add_test(tc_core, test_vunit10);
    tcase_add_test(tc_core, test_issue942);
+   tcase_add_test(tc_core, test_lcs2016_i03);
    suite_add_tcase(s, tc_core);
 
    return s;