+++ NEWS.md
@@ -16,6 +16,8 @@
 - The `--force-init` command which was deprecated in the 1.7 release has
   been removed.
 - Coverage exclude files now support `fold` command.
+- A signature is now allowed in the formal part of generic map
+  associations in VHDL-2019 mode.
 
 ## Version 1.13.2 - 2024-08-11
 - Fixed an incorrect bounds check error when a constant declaration has